% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLenDistComm.R
\name{boxplotLenDistComm}
\alias{boxplotLenDistComm}
\title{BoxPlot of Commercial Length Data}
\usage{
boxplotLenDistComm(dat, meditsCode, GSA)
}
\arguments{
\item{dat}{data.frame of Fisheries landings at length data}

\item{meditsCode}{medits code}
}
\description{
BoxPlot of Commercial Length Data
}
\details{
Box plot of length distributions of the commercial landings aggregated across metiers.
}
