% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchSummary.R
\name{catchSummary}
\alias{catchSummary}
\title{Summary table for a focal Species and GSA}
\usage{
catchSummary(con, code = "ELEDMOS", index = "ind_skm", stdat, gsa = "all",
  matFrac = FALSE, matStage = 3)
}
\arguments{
\item{con}{connection to the MEDITS database.}

\item{code}{species code}

\item{index}{index of abundance of interest. It can be ind_skm or p_skm for biomass}

\item{stdat}{data.frame with data on strata}

\item{gsa}{gsa number or a vector of numbers (e.g. c(1,5,6)), or "all".}

\item{matFrac}{whether to correct the index for the fraction of fish having a certain maturation stage.}

\item{matStage}{MEDITS maturity code, from 0 to 4 and ND. 0 undetermined, 1 = immature, 2 maturing, 3 mature, 4 spent resting. See manual for details.}

\item{dat}{specific data with catches}
}
\description{
Summary table for a focal Species and GSA
}
