% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectMEDITS.R
\name{connectMedits}
\alias{connectMedits}
\title{Create Connection to MEDITS Database on localhost.}
\usage{
connectMedits(dbuser = "postgres")
}
\arguments{
\item{dbuser}{role name for database}
}
\description{
Create Connection to MEDITS Database on localhost.
}
