% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corPlot.R
\name{corPlot}
\alias{corPlot}
\title{Correlation Plot of Catch at Age}
\usage{
corPlot(dat, species, GSA)
}
\arguments{
\item{species}{as MEDITS code}

\item{agedat}{data.frame having all age class data by species, GSA and year}
}
\description{
Correlation Plot of Catch at Age
}
