% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatLat.R
\name{formatLat}
\alias{formatLat}
\title{Format MEDITS Coordinates}
\usage{
formatLat(latmed, lonmed, quadrant)
}
\arguments{
\item{latmed}{latitude in MEDITS format}

\item{lonmed}{longitude in MEDITS format}

\item{quadrant}{code for geographic quadrant. It could be 1 for hauls east of longitude 0, and 7 for hauls west of longitude 0.}
}
\description{
It transorms the coordinates of the hauls in decimal degrees.
}
