% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnualSamples.R
\name{getAnnualSamples}
\alias{getAnnualSamples}
\title{Get Annual Samples}
\usage{
getAnnualSamples(dat = getFishLandData())
}
\arguments{
\item{dat}{data.frame from \code{getFishLandData()} or \code{getDiscLenData()}}
}
\description{
Used to get the number of measured animals each year for all species in all GSA
}
