% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDiscLandData.R
\name{getDiscLenData}
\alias{getDiscLenData}
\title{Get and Process Fisheries Discards at Length Data}
\usage{
getDiscLenData(dataFile = "../data/DCF_catch/discards_length.csv")
}
\arguments{
\item{dataFile}{csv of the Fisheries Landings at Length Data}
}
\description{
Get and Process Fisheries Discards at Length Data
}
