% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFishLandData.R
\name{getFishLandData}
\alias{getFishLandData}
\title{Get and Process Fisheries Landings at Length Data}
\usage{
getFishLandData(dataFile = "../data/DCF_catch/landings_length.csv")
}
\arguments{
\item{dataFile}{csv of the Fisheries Landings at Length Data}
}
\description{
Get and Process Fisheries Landings at Length Data
}
