% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqMatSummary.R
\name{getFreqData}
\alias{getFreqData}
\title{Get Maturity Specific CPUE}
\usage{
getFreqData(con, code = "ELEDMOS", gsa = 11, stdat, matStage = 3)
}
\arguments{
\item{con}{connection to the medits database}

\item{code}{species code}

\item{gsa}{gsa number or a vector of numbers (e.g. c(1,5,6)), or "all".}

\item{stdat}{data.frame with data on strata}

\item{matStage}{MEDITS maturity code, from 0 to 4 and ND. 0 undetermined, 1 = immature, 2 maturing, 3 mature, 4 spent resting. See manual for details.}
}
\description{
This function returns the medits cpue in terms of maturity stage
}
