% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGLMfit.R
\name{getGLMfit}
\alias{getGLMfit}
\title{Get Standardized CPUE Through GLM Fit}
\usage{
getGLMfit(con, code = "ELEDMOS", index = "ind_skm", stdat, gsa = "all")
}
\arguments{
\item{con}{connection to the medits database}

\item{code}{species code}

\item{index}{index of abundance of interest. It can be ind_skm or p_skm for biomass}

\item{stdat}{data.frame with data on strata}

\item{gsa}{gsa number either a number, a vector of numbers or "all".}
}
\value{
a data.frame with predicted standardized CPUES. For The moment only the model for individuals is implemented as the biomass model should use a delta-LN or delta-Gamma model. This should be implemented shortly.
}
\description{
Get Standardized CPUE Through GLM Fit
}
