% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpecData.R
\name{getSpecData}
\alias{getSpecData}
\title{Get MEDITS data by Species}
\usage{
getSpecData(con, code = "ELEDMOS", gsa = "all")
}
\arguments{
\item{con}{connection to the MEDITS database.}

\item{code}{species code}

\item{gsa}{gsa number or a vector of numbers (e.g. c(1,5,6)), or "all".}
}
\description{
Get MEDITS data by Species
}
