% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStockScore.R
\name{getStockScore}
\alias{getStockScore}
\title{Get Score for Stock}
\usage{
getStockScore(dat = tab1, code = "ARISFOL", GSA = 9)
}
\arguments{
\item{dat}{data.frame having number of individuals per GSA, Stock and year.}

\item{code}{MEDITS code}

\item{GSA}{MEDITS gsa}
}
\description{
Get Score for Stock
}
