% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSumStats.R
\name{getSumStats}
\alias{getSumStats}
\title{Get Summary Statistics for Index Time Series}
\usage{
getSumStats(con, code, gsa, value = F)
}
\arguments{
\item{con}{connection to the MEDITS database}

\item{code}{species code}

\item{gsa}{gsa number, vector of gsas (for groups) or "all"}

\item{value}{logical if TRUE return the data table.}
}
\description{
Get Summary Statistics for Index Time Series
}
