% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTA.R
\name{getTA}
\alias{getTA}
\title{Load TA from the Database}
\usage{
getTA(con)
}
\arguments{
\item{con}{a connection to the medits Database.}
}
\description{
It loads the TA table and preforms some data management to obtain desired fields for analysis
}
