% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplotCS.R
\name{multiplotCS}
\alias{multiplotCS}
\title{Multipanel Plot for Catch Summary Page}
\usage{
multiplotCS(con, code, gsa)
}
\arguments{
\item{con}{connection to database}

\item{code}{species code}

\item{gsa}{number, vector of numbers (for groups) or "all"}
}
\description{
It produces the multipanel figure for the catch summary page as a pdf file.
}
