% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplotCS.R
\name{multiplotSM}
\alias{multiplotSM}
\title{Multipanel Plot for Size and Maturity Stats}
\usage{
multiplotSM(con, code, gsa)
}
\arguments{
\item{con}{connection to database}

\item{code}{species code}

\item{gsa}{number, vector of numbers (for groups) or "all"}
}
\description{
It produces the multipanel figure for the size and maturity summary statistics.
}
