% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgeTrend.R
\name{plotAgeTrend}
\alias{plotAgeTrend}
\title{Plot Age Class Time Series}
\usage{
plotAgeTrend(agedat, species, GSA)
}
\arguments{
\item{agedat}{data.frame having all age class data by species, GSA and year}
}
\description{
Plot Age Class Time Series
}
