% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAgeTrend.R
\name{plotAgeTrendFL}
\alias{plotAgeTrendFL}
\title{Cross Correlation Matrix of Catch at Age}
\usage{
plotAgeTrendFL(dat, species, GSA)
}
\arguments{
\item{species}{as MEDITS code}

\item{agedat}{data.frame having all age class data by species, GSA and year}
}
\description{
Cross Correlation Matrix of Catch at Age
}
