% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCatchDateDis.R
\name{plotCatchDateDis}
\alias{plotCatchDateDis}
\title{Plot Catch Date Distribution}
\usage{
plotCatchDateDis(con, code, gsa)
}
\arguments{
\item{con}{connection to the MEDITS database.}

\item{code}{species code.}

\item{gsa}{gsa number, vector of numbers (for gsa groups), or "all" for the entire MEDITS' survey area.}
}
\description{
boxplot showing min, max and mean day in year of survey data by year (or 5,50,95%).
}
