% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCatchSummary.R
\name{plotCatchSummary}
\alias{plotCatchSummary}
\title{Plot Index of Abundance Over Time}
\usage{
plotCatchSummary(dat = dat1, ylab = "#/km2")
}
\arguments{
\item{dat}{data.frame coming from \code{catchSummary}.}

\item{ylab}{lab for y-axis in plot}
}
\description{
Plot Index of Abundance Over Time
}
