% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDepthDis.R
\name{plotDepthDis}
\alias{plotDepthDis}
\title{Plot Depth Distribution Density}
\usage{
plotDepthDis(dat)
}
\arguments{
\item{dat}{a merged TA and TB file with catches of a given species. This data.frame can be produced with the function \code{getSpecData()}}
}
\description{
Plot Depth Distribution Density
}
