% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLenDistComm.R
\name{plotLenDistComm}
\alias{plotLenDistComm}
\title{Plot Length Distribution Commercial}
\usage{
plotLenDistComm(dat, meditsCode, GSA)
}
\arguments{
\item{dat}{data.frame of Fisheries landings at length data}

\item{commCode}{commercial code}
}
\description{
Plot Length Distribution Commercial
}
\details{
This function shows the contribution of each metier as stacked bar. Showing a plot for each metier was not practical as it would have generated tens of panels for species and GSA.
}
