% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measuredAnimals.R
\name{plotMeasuredAnimals}
\alias{plotMeasuredAnimals}
\title{Plot Measured Animals}
\usage{
plotMeasuredAnimals(dat = tab1, code)
}
\arguments{
\item{dat}{data.frame of aggregated length or discard data by GSA and Species}

\item{code}{MEDITS code}
}
\description{
Plot Measured Animals
}
