% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableLenDis.R
\name{tableLenDis}
\alias{tableLenDis}
\title{Tabulate Length Distribution}
\usage{
tableLenDis(con, code, gsa)
}
\arguments{
\item{con}{database connection}

\item{code}{species code}

\item{gsa}{gsa number or a vector of numbers (e.g. c(1,5,6)), or "all".}
}
\description{
it connects to the tc table and estract the focal species and gsa. Then it tabulates the length classes of the sample
}
