% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measuredAnimals.R
\name{tabMeasuredAnimals}
\alias{tabMeasuredAnimals}
\title{Tabulate Measured Animals}
\usage{
tabMeasuredAnimals(dat = tab1, code)
}
\arguments{
\item{dat}{data.frame of aggregated length or discard data by GSA and Species}

\item{code}{MEDITS code}
}
\description{
Tabulate Measured Animals
}
