% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSpage.R
\name{writeCSpage}
\alias{writeCSpage}
\title{Write Catch Summary Page}
\usage{
writeCSpage(code, gsa, sciname)
}
\arguments{
\item{code}{species code}

\item{gsa}{number, vector of numbers (for gsa groups), or "all" for the entire survey domain.}

\item{sciname}{scientific name of the species.}
}
\description{
This function writes the catch summary page for each GSA. It writes in the \code{tex/} folder. It generates a Rnw file named \code{catchSummaryPage[code][gsa].Rnw}.
}
