% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSpage.R
\name{writeGSApage}
\alias{writeGSApage}
\title{Write GSA summary Page}
\usage{
writeGSApage(gsa)
}
\arguments{
\item{gsa}{number, vector of numbers (for gsa groups), or "all" for the entire survey domain.}
}
\description{
This function writes a GSA summary page that precedes the species catch summaries for that GSA. It writes into the \code{tex} folder.
}
