% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCSpage.R
\name{writeSI}
\alias{writeSI}
\title{Write SweaveInputs for CS Pages}
\usage{
writeSI(gsa)
}
\arguments{
\item{gsa}{number, vector of numbers (for gsa groups), or "all" for the entire survey domain.}
}
\value{
Include only the species summaries that have correspondent 6-panel figures. This is a good check to identify data deficient species.
}
\description{
This functions list the species summary pages as SweaveInputs in a Rnw file called \code{gsa[n]summaries.Rnw} placed into the \code{tex} folder.
}
